requirements.txt
# ── Core NLP ──────────────────────────────────────────────
pdfplumber>=0.10
scikit-learn>=1.3
numpy>=1.24
sentence-transformers>=2.2      # embeddings (requiere HuggingFace en 1ra ejecución)

# ── Índice vectorial ──────────────────────────────────────
faiss-cpu>=1.7                  # búsqueda vectorial eficiente (IndexFlatIP)

# ── QA Transformer / RAG ─────────────────────────────────
transformers>=4.35              # DistilBERT QA pipeline
torch>=2.0                      # backend (CPU suficiente para inferencia)

# Modelo español: mrm8488/distill-bert-base-spanish-wwm-cased-finetuned-spa-squad2-es
# Descargar una vez con: python -c "from transformers import pipeline; pipeline('question-answering', model='mrm8488/distill-bert-base-spanish-wwm-cased-finetuned-spa-squad2-es')"

# ── API ───────────────────────────────────────────────────
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0

# ── Visualización ─────────────────────────────────────────
matplotlib>=3.7
seaborn>=0.12
pandas>=2.0

# ── Tests ─────────────────────────────────────────────────
httpx>=0.26
